@@ -6,8 +6,8 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build:github": "ng build -e=prod -prod --base-href=/amazing-time-picker --deploy-url=/amazing-time-picker/ --progress false",
-    "build": "ng build -e=prod -prod  --progress false",
+    "build:github": "ng build demo --base-href=/amazing-time-picker --deploy-url=/amazing-time-picker/ --progress false",
+    "build": "ng build amazing-time-picker --prod --progress false",
     "test": "ng test",
     "lint": "ng lint",
     "ci": "npm run serve & npm run serve:beds:5 & npm run serve:beds:4 & npm run serve:beds:2 & npm run cypress:ci",
@@ -15,35 +15,37 @@
     "serve:beds:4": "PORT=12001 ./node_modules/.bin/serve /tmp/amazing-time-picker-test-beds/ng4-test-bed/dist",
     "serve:beds:5": "PORT=12002 ./node_modules/.bin/serve /tmp/amazing-time-picker-test-beds/ng5-test-bed/dist",
     "cypress:ci": "./node_modules/.bin/cypress run ",
-    "serve": "PORT=8080 ./node_modules/.bin/serve dist",
+    "serve": "ng build demo && PORT=8080 ./node_modules/.bin/serve dist/demo",
     "e2e": "ng e2e",
     "cypress": "./node_modules/.bin/cypress open",
     "hmr": "ng serve --hmr -e=hmr --port=8080",
-    "packagr": "ng-packagr -p ng-package.json",
-    "gh-pages": "git push origin :gh-pages && git subtree push --prefix dist origin gh-pages",
+    "gh-pages": "ng build demo && git push origin :gh-pages && git subtree push --prefix dist/demo origin gh-pages",
     "build:npm": "npm run publisher",
-    "publisher": "rm npm_dist -Rf && npm run packagr && cd npm_dist && mv $(npm pack) dist.tgz"
+    "publisher": "ng build amazing-time-picker --prod && npm publish dist/amazing-time-picker"
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": ">=4.2.4",
-    "@angular/common": ">=2.4.5",
-    "@angular/compiler": ">=2.4.5",
-    "@angular/core": ">=2.4.5",
-    "@angular/forms": ">=2.4.5",
-    "@angular/http": ">=2.4.5",
-    "@angular/platform-browser": ">=2.4.5",
-    "@angular/platform-browser-dynamic": ">=2.4.5",
-    "@angular/router": ">=2.4.5",
+    "@angular/animations": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
     "core-js": "^2.4.1",
     "intl": "^1.2.5",
-    "rxjs": "^5.4.2",
+    "rxjs": "6.1.0",
+    "rxjs-compat": "^6.0.0-rc.0",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.2",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular-devkit/build-ng-packagr": "~0.6.0",
+    "@angular/cli": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/language-service": "6.0.0",
     "@angularclass/hmr": "^2.1.3",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
@@ -59,11 +61,13 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^1.6.0",
+    "ng-packagr": "^2.4.2",
     "protractor": "~5.1.2",
     "serve": "^6.4.8",
     "ts-node": "~3.2.0",
+    "tsickle": ">=0.25.5",
+    "tslib": "^1.7.1",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "2.7.2"
   }
 }
