@@ -2,7 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { appRoutersGenerator } from './app.router';
 
-import { AmazingTimePickerModule } from './atp-library/atp-time-picker.module';
+import { AmazingTimePickerModule } from 'amazing-time-picker';
 import { AppComponent } from './app.component';
 import { TabsComponent } from './presentation/tabs/tabs.component';
 import { TabComponent } from './presentation/tab/tab.component';
